+++ src/components/menus/MenuButton.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 
-import { makeStyles, Menu, MenuItem } from '@material-ui/core';
+import { Icon, makeStyles, Menu, MenuItem } from '@material-ui/core';
 
 import { Button } from 'custom-components';
 
@@ -10,14 +10,31 @@ import { authRoles } from 'auth';
 const useStyles = makeStyles((theme) => ({
     root: {},
     dropdown: {
-        top: '56px!important'
+        top: '65px!important',
+        minWidth: '150px',
+        '& a, & div': {
+            display: 'flex',
+            alignItems: 'center',
+            padding: '0 15px',
+            '& li:hover': {
+                backgroundColor: theme.palette.background.default
+            },
+            '&:hover': {
+                cursor: 'pointer',
+                backgroundColor: theme.palette.background.default
+            }
+        }
     },
     link: {
         color: theme.palette.text.primary
+    },
+    itemIcon: {
+        opacity: 0.6,
+        marginRight: '5px'
     }
 }));
 
-function MenuButton({ children, items, role }) {
+function MenuButton({ children: title, component: Component = Button, items, onLogout, role }) {
     const classes = useStyles();
 
     const [anchorEl, setAnchorEl] = useState(null);
@@ -31,20 +48,30 @@ function MenuButton({ children, items, role }) {
     };
 
     const open = Boolean(anchorEl);
-    const listItems = items?.map((link) => {
+    const listItems = items?.map((link, index) => {
         if (link.roles && !link.roles?.includes(authRoles[role])) return null;
-        return (
-            <Link key={link.path} className={classes.link} to={link.path}>
-                <MenuItem onClick={handleClose}>{link.title}</MenuItem>
-            </Link>
-        );
+
+        if (onLogout && link.type === 'logout')
+            return (
+                <div key={index}>
+                    {link.icon && <Icon className={classes.itemIcon}>{link.icon}</Icon>}
+                    <MenuItem onClick={onLogout}>{link.title}</MenuItem>
+                </div>
+            );
+        else
+            return (
+                <Link key={index} className={classes.link} to={link.path}>
+                    {link.icon && <Icon className={classes.itemIcon}>{link.icon}</Icon>}
+                    <MenuItem onClick={handleClose}>{link.title}</MenuItem>
+                </Link>
+            );
     });
 
     return (
         <div className={classes.root}>
-            <Button onClick={handleMenu} variant="text">
-                {children}
-            </Button>
+            <Component onClick={handleMenu} variant="text">
+                {title}
+            </Component>
             <Menu
                 classes={{ paper: classes.dropdown }}
                 anchorEl={anchorEl}